refactor: Rename API Gateway to Edge
