+++ edge/pom.xml
@@ -10,9 +10,9 @@
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.github.itsAkshayDubey.event-driven-architecture</groupId>
-	<artifactId>api-gateway</artifactId>
+	<artifactId>edge</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>api-gateway</name>
+	<name>edge</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
 		<java.version>1.8</java.version>
