+++ edge/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/edge/Edge.java
@@ -1,4 +1,4 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.apigateway;
+package com.github.itsAkshayDubey.eventdrivenarchitecture.edge;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 
 @SpringBootApplication
 @EnableEurekaClient
-public class ApiGatewayApplication {
+public class Edge {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ApiGatewayApplication.class, args);
+		SpringApplication.run(Edge.class, args);
 	}
 
 }
