+++ edge/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 server.port=8082
-spring.application.name=api-gateway
+spring.application.name=edge
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka
 spring.cloud.gateway.discovery.locator.enabled: true
 spring.cloud.gateway.discovery.locator.lower-case-service-id= true
