+++ edge/src/test/java/com/github/itsAkshayDubey/eventdrivenarchitecture/edge/EdgeTests.java
@@ -1,10 +1,10 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.apigateway;
+package com.github.itsAkshayDubey.eventdrivenarchitecture.edge;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApiGatewayApplicationTests {
+class EdgeTests {
 
 	@Test
 	void contextLoads() {
