+++ edge/stop.sh
